+++ jgclark.Summaries/src/saveSearch.js
@@ -12,6 +12,7 @@
 import {
   displayTitle,
   stringReplace,
+  getFolderFromFilename,
 } from '../../helpers/general'
 import {
   showMessage,
@@ -28,7 +29,8 @@ import {
   monthNameAbbrev,
   withinDateRange,
   dateStringFromCalendarFilename,
-  toLocaleShortTime,
+  // toLocaleShortTime,
+  nowLocaleDateTime,
 } from '../../helpers/dateTime'
 import { getOrMakeConfigurationSection } from '../../nmn.Templates/src/configuration'
 import {
@@ -41,11 +43,9 @@ import {
 // Config settings
 // Globals, to be looked up later
 let pref_folderToStore: string
-let pref_occurrencesHeading: string
+let pref_foldersToIgnore: Array<string> = []
 let pref_headingLevel: 1 | 2 | 3 | 4 | 5
-let pref_occurrencesToMatch: $ReadOnlyArray<string> = []
 let pref_highlightOccurrences: boolean
-let pref_showEmptyOccurrences: boolean
 let pref_addDates: string
 
 async function getPluginSettings(): Promise<void> {
@@ -59,8 +59,8 @@ async function getPluginSettings(): Promise<void> {
     console.log("\tCouldn't find 'summaries' settings in _configuration note.")
     return
   }
-  console.log("\tFound 'summaries' settings in _configuration note.")
 
+  console.log("\tFound 'summaries' settings in _configuration note.")
   // now get each setting
   pref_folderToStore =
     summConfig.folderToStore != null
@@ -68,25 +68,18 @@ async function getPluginSettings(): Promise<void> {
       ? summConfig.folderToStore
       : 'Summaries'
   // console.log(pref_folderToStore)
-  pref_occurrencesHeading =
-    summConfig.occurrencesHeading != null
+  pref_foldersToIgnore =
+    summConfig.foldersToIgnore != null
       // $FlowIgnore[incompatible-type]
-      ? summConfig.occurrencesHeading
-      : 'Occurrences'
-  // console.log(pref_occurrencesHeading)
+      ? summConfig.foldersToIgnore
+      : ['📋 Templates']
+  // console.log(String(pref_foldersToIgnore))
   pref_headingLevel =
     summConfig.headingLevel != null
       // $FlowIgnore[incompatible-type]
       ? summConfig.headingLevel
       : 2
   // console.log(pref_headingLevel)
-  pref_occurrencesToMatch =
-    // $FlowIgnore[incompatible-type]
-    summConfig.occurrencesToMatch != null
-      // $FlowIgnore[incompatible-type]
-      ? summConfig.occurrencesToMatch
-      : []
-  // console.log(pref_occurrencesToMatch)
   pref_highlightOccurrences =
     summConfig.highlightOccurrences != null
       // $FlowIgnore[incompatible-type]
@@ -103,75 +96,52 @@ async function getPluginSettings(): Promise<void> {
 
 //-------------------------------------------------------------------------------
 // Ask user which period to cover, call main stats function, and present results
-export async function occurrencesPeriod(): Promise<void> {
-  // Get config settings from Template folder _configuration note
+export async function saveSearch(): Promise<void> {
+  // get relevant settings
   await getPluginSettings()
 
-  // Generate main data
-  const [fromDate, toDate, periodString, periodPartStr] = await getPeriodStartEndDates()
-
-  if (fromDate == null || toDate == null) {
-    console.log('dates could not be parsed')
-    return
-  }
-  const fromDateStr = unhyphenatedDate(fromDate) //fromDate.toISOString().slice(0, 10).replace(/-/g, '')
-  const toDateStr = unhyphenatedDate(toDate) // toDate.toISOString().slice(0, 10).replace(/-/g, '')
-  
-  // If a single search term was given, use that, otherwise use the setting pref_occurrencesToMatch
-  // const stringsToMatch = Array.from(pref_occurrencesToMatch)
-  const stringsToMatch = Array.from(pref_occurrencesToMatch)
-  console.log(
-    `\nperiodOccurrences: looking for '${String(stringsToMatch)}' over ${periodString} (${fromDateStr}-${toDateStr}):`,
-  )
-
-  // Get array of all daily notes that are within this time period
-  const periodDailyNotes = DataStore.calendarNotes.filter((p) =>
-    withinDateRange(
-      dateStringFromCalendarFilename(p.filename),
-      fromDateStr,
-      toDateStr,
-    ),
-  )
-  if (periodDailyNotes.length === 0) {
-    console.log('  warning: no matching daily notes found')
-    await showMessage(`No matching daily notes found; stopping.`)
-    return
-  }
-
-  // Find matches in notes for the time period
-  const outputArray = []
-  for (const toMatch of stringsToMatch) {
-    // output a heading first
-    // get list of matching paragraphs for this string
-    const results = await gatherMatchingLinesInPeriod(periodDailyNotes, toMatch)
-    const lines = results?.[0]
-    const dates = results?.[1]
-    if (lines.length > 0) {
-      console.log(`  Found ${lines.length} results for ${toMatch}`)
-      outputArray.push(`### ${toMatch}`)
-      // format the output
-      for (let i = 0; i < lines.length; i++) {
-        // datePart depends on pref_addDates setting
-        const datePart = (pref_addDates === 'links')
-          ? ` >${toISODateString(dates[i])}`
-          : (pref_addDates === 'dates')
-            ? ` (${toLocaleDateString(dates[i])})`
-            : ''
-        outputArray.push(`- ${lines[i]}${datePart}`)
-      }
-    } else if (pref_showEmptyOccurrences) {
-      // If there's nothing to report, make that clear
-      outputArray.push(`### ${toMatch}`)
-      outputArray.push('(none)')
+  // Create list of project notes not in excluded folders
+  const allProjectNotes = DataStore.projectNotes
+  const projectNotesToInclude = []
+  // Iterate over the folders ...
+  for (const pn of allProjectNotes) {
+    const thisFolder = getFolderFromFilename(pn.filename)
+    if (!pref_foldersToIgnore.includes(thisFolder)) {
+      projectNotesToInclude.push(pn)
+    } else {
+      console.log(pn.filename)
     }
+  }
+  console.log(`Will use ${projectNotesToInclude.length} project notes out of ${allProjectNotes.length}`)
+  // Add all the calendar notes
+  const notes = DataStore.calendarNotes.concat(projectNotesToInclude)
+
+  // Ask user for search term
+  const searchTerm = await getInput(`Exact word/phrase to search for`)
+
+  // Find matches in this set of notes for the time period
+  const outputArray = []
+  // output a heading first
+  const results = await gatherMatchingLines(notes, searchTerm)
+  const lines = results?.[0]
+  const contexts = results?.[1]
+  if (lines.length > 0) {
+    // outputArray.push(`### ${searchTerm}`)
+    console.log(`  Found ${lines.length} results for '${searchTerm}'`)
+    // format the output
+    for (let i = 0; i < lines.length; i++) {
+      outputArray.push(`- ${lines[i]} ${contexts[i]}`)
+    }
+  } else {
+    // If there's nothing to report, make that clear
+    // outputArray.push(`### ${searchTerm}`)
+    outputArray.push('(no matches)')
   }
 
   // Ask where to save this summary to
-  const labelString = `🖊 Create/update note '${periodString}' in folder '${String(
-    pref_folderToStore,
-  )}'`
+  const labelString = `🖊 Create/update note in folder '${pref_folderToStore}'`
   const destination = await chooseOption(
-    `Where to save the summary for ${periodString}?`,
+    `Where should I save the search results?`,
     [
       {
         // TODO: When weekly/monthly notes are made possible in NP, then add options like this
@@ -198,6 +168,8 @@ export async function occurrencesPeriod(): Promise<void> {
   )
 
   // Ask where to send the results
+  const currentDate = nowLocaleDateTime
+  const headingString = `Search results for '${searchTerm}'`
   switch (destination) {
     case 'current': {
       const currentNote = Editor.note
@@ -209,7 +181,7 @@ export async function occurrencesPeriod(): Promise<void> {
         )
         const insertionLineIndex = currentNote.paragraphs.length
         currentNote.insertHeading(
-          `${pref_occurrencesHeading} ${periodPartStr}`,
+          `${headingString} at ${currentDate}`,
           insertionLineIndex,
           pref_headingLevel,
         )
@@ -222,14 +194,16 @@ export async function occurrencesPeriod(): Promise<void> {
       break
     }
     case 'note': {
+      const requestedTitle = await getInput(`What do you want to call this note?`)
+
       let note: TNote
       // first see if this note has already been created
       // (look only in active notes, not Archive or Trash)
       const existingNotes: $ReadOnlyArray<TNote> =
-        DataStore.projectNoteByTitle(periodString, true, false) ?? []
+        DataStore.projectNoteByTitle(requestedTitle, true, false) ?? []
 
       console.log(
-        `\tfound ${existingNotes.length} existing summary notes for this period`,
+        `\tfound ${existingNotes.length} existing ${requestedTitle} notes`,
       )
 
       if (existingNotes.length > 0) {
@@ -237,9 +211,9 @@ export async function occurrencesPeriod(): Promise<void> {
         // console.log(`\tfilename of first matching note: ${displayTitle(note)}`)
       } else {
         // make a new note for this. NB: filename here = folder + filename
-        const noteFilename = DataStore.newNote(periodString, pref_folderToStore) ?? ''
-        if (!noteFilename) {
-          console.log(`\tError creating new note (filename: ${noteFilename})`)
+        const noteFilename = DataStore.newNote(requestedTitle, pref_folderToStore) ?? ''
+        if (noteFilename === '') {
+          console.log(`\tError creating new note (filename '${noteFilename}')`)
           await showMessage('There was an error creating the new note')
           return
         }
@@ -254,16 +228,17 @@ export async function occurrencesPeriod(): Promise<void> {
       }
       console.log(`\twriting results to the new note '${displayTitle(note)}'`)
 
+      // const note = note
       // Do we have an existing Hashtag counts section? If so, delete it.
       const insertionLineIndex = removeSection(
         note,
-        pref_occurrencesHeading,
+        headingString,
       )
       console.log(`\tinsertionLineIndex: ${String(insertionLineIndex)}`)
       // Set place to insert either after the found section heading, or at end of note
       // write in reverse order to avoid having to calculate insertion point again
       note.insertHeading(
-        `${pref_occurrencesHeading} ${periodPartStr}`,
+        `${headingString} at ${currentDate}`,
         insertionLineIndex,
         pref_headingLevel,
       )
@@ -274,13 +249,13 @@ export async function occurrencesPeriod(): Promise<void> {
       )
       Editor.openNoteByFilename(note.filename)
 
-      console.log(`\twritten results to note '${periodString}'`)
+      console.log(`\twritten results to note '${requestedTitle}'`)
       break
     }
 
     case 'log': {
       console.log(
-        `Summaries for ${periodString} ${periodPartStr}`,
+        `Search results for '${searchTerm}', ${currentDate}`,
       )
       console.log(outputArray.join('\n'))
       break
@@ -293,34 +268,35 @@ export async function occurrencesPeriod(): Promise<void> {
 }
 
 //-------------------------------------------------------------------------------
-// Return list of lines matching the specified string in daily notes of a given time period.
-// @param {string} periodDailyNotes - array of Notes to look over
+// Return list of lines matching the specified string in project or daily notes.
+// @param {string} notes - array of Notes to look over
 // @param {string} stringToLookFor - string to look for
-// @return [Array, Array] - array of lines with matching hashtag, and array of dates of those lines
-async function gatherMatchingLinesInPeriod(
-  periodDailyNotes,
+// @return [Array, Array] - array of lines with matching hashtag, and array of 
+//   contexts for those lines (dates for daily notes; title for project notes).
+async function gatherMatchingLines(
+  notes,
   stringToLookFor
-): Promise<[Array<string>, Array<Date>]> {
-  console.log(`Looking for '${stringToLookFor}' in ${periodDailyNotes.length} notes`)
-  CommandBar.showLoading(true, `Searching in ${periodDailyNotes.length} notes ...`)
+): Promise<[Array<string>, Array<string>]> {
+
+  console.log(`Looking for '${stringToLookFor}' in ${notes.length} notes`)
+  CommandBar.showLoading(true, `Searching in ${notes.length} notes ...`)
   await CommandBar.onAsyncThread()
 
-  // work out what set of mentions to look for (or ignore)
-  // console.log(`Looking for '${stringToLookFor}' in ${periodDailyNotes.length} relevant daily notes`)
-
-  const matches = []
-  const dates: Array<Date> = []
+  const matches: Array<string> = []
+  const noteContexts: Array<string> = []
   let i = 0
-  for (const n of periodDailyNotes) {
+  for (const n of notes) {
     i += 1
-    const noteDate = n.date
+    const noteContext = (n.date != null)
+      ? `>${toISODateString(n.date)}`
+      : `[[${n.title ?? ''}]]`
     // find any matches
     const matchingParas = n.paragraphs.filter((q) => q.content.includes(stringToLookFor))
     for (const p of matchingParas) {
       // If the stringToLookFor is in the form of an 'attribute::' and found at the start of a line,
       // then remove it from the output line
       let matchLine = p.content
-      console.log(`  Found '${stringToLookFor}' in ${matchLine} (${String(noteDate)})`)
+      // console.log(`  Found '${stringToLookFor}' in ${matchLine} (${noteContext})`)
       if (stringToLookFor.endsWith('::') && matchLine.startsWith(stringToLookFor)) {
         matchLine = matchLine.replace(stringToLookFor, '') // NB: only removes first instance
         // console.log(`    -> ${matchLine}`)
@@ -331,14 +307,13 @@ async function gatherMatchingLinesInPeriod(
       }
       // console.log(`    -> ${matchLine}`)
       matches.push(matchLine.trim())
-      // $FlowFixMe[incompatible-call]
-      dates.push(noteDate)
+      noteContexts.push(noteContext)
     }
-    if (i % 10 === 0) {
-      CommandBar.showLoading(true, `Searching in ${periodDailyNotes.length} notes ...`, (i / periodDailyNotes.length))
+    if (i % 20 === 0) {
+      CommandBar.showLoading(true, `Searching in ${notes.length} notes ...`, (i / notes.length))
     }
   }
   await CommandBar.onMainThread()
   CommandBar.showLoading(false)
-  return [matches, dates]
+  return [matches, noteContexts]
 }